perbaikan fitur relational delete

+++ app/Http/Controllers/DonorDarahEventController.php
@@ -146,7 +146,10 @@ class DonorDarahEventController extends Controller
      */
     public function destroy(DonorDarahEvent $donorDarahEvent)
     {
-        $donorDarahEvent->registrants()->delete();
+        if ($donorDarahEvent->registrants) {
+            $donorDarahEvent->registrants()->delete();
+        }
+
         $donorDarahEvent->delete();
 
         return redirect()->route('donor-darah-events')->with('success', 'Donor Darah Event deleted successfully.');
